+++ kinto_http/tests/support.py
@@ -46,7 +46,7 @@ def get_200():
     response_200 = mock.MagicMock()
     response_200.status_code = 200
     response_200.json().return_value = mock.sentinel.resp,
-    response_200.headers = mock.sentinel.headers
+    response_200.headers = {}
     return response_200
 
 
@@ -70,5 +70,6 @@ def get_503():
 
 def get_403():
     response_403 = mock.MagicMock()
+    response_403.headers = {}
     response_403.status_code = 403
     return response_403